+++ app/Livewire/Settings/Tabs/BudgetGroupManager.php
@@ -20,7 +20,7 @@ class BudgetGroupManager extends Component
 
     public ?string $description = null;
 
-    #[Validate('nullable|numeric|min:0|max:100')]
+    #[Validate('nullable|integer|min:0|max:100')]
     public ?string $percentage = null;
 
     public ?int $editingId = null;
@@ -52,7 +52,7 @@ class BudgetGroupManager extends Component
         $totalPercentage = $this->calculateTotalPercentage();
 
         if ($totalPercentage > 100) {
-            $this->addError('percentage', 'Total persentase semua kelompok anggaran tidak boleh melebihi 100%. Saat ini: '.number_format($totalPercentage, 2).'%');
+            $this->addError('percentage', 'Total persentase semua kelompok anggaran tidak boleh melebihi 100%. Saat ini: ' . number_format($totalPercentage, 2) . '%');
 
             return;
         }
@@ -61,7 +61,7 @@ class BudgetGroupManager extends Component
             'code' => $this->code,
             'name' => $this->name,
             'description' => $this->description,
-            'percentage' => $this->percentage ? (float) $this->percentage : null,
+            'percentage' => $this->percentage ? (int) $this->percentage : null,
         ];
 
         if ($this->editingId) {
@@ -86,7 +86,7 @@ class BudgetGroupManager extends Component
         $this->code = $budgetGroup->code;
         $this->name = $budgetGroup->name;
         $this->description = $budgetGroup->description;
-        $this->percentage = $budgetGroup->percentage ? (string) $budgetGroup->percentage : null;
+        $this->percentage = $budgetGroup->percentage ? (string) (int) $budgetGroup->percentage : null;
         $this->modalTitle = 'Edit Kelompok Anggaran';
         $this->dispatch('open-modal', modalId: 'modal-budget-group');
     }
